+++ practical_3/player.h
@@ -7,7 +7,7 @@ private:
 	float _speed;
 
 public:
-	void update(double dt) override;
+	void update(float dt) override;
 	Player();
 	void render(sf::RenderWindow &window) const override;
 };
